+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -88,6 +88,7 @@ from snowflake.snowpark.functions import (
     count_distinct,
     date_part,
     date_trunc,
+    dateadd,
     dayofmonth,
     dayofyear,
     dense_rank,
@@ -10073,6 +10074,15 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             "dayofweek": (lambda column: builtin("dayofweekiso")(col(column)) - 1),
             "microsecond": (lambda column: floor(date_part("ns", col(column)) / 1000)),
             "nanosecond": (lambda column: date_part("ns", col(column)) % 1000),
+            "is_month_start": (
+                lambda column: coalesce(dayofmonth(col(column)) == 1, pandas_lit(False))
+            ),
+            "is_month_end": (
+                lambda column: coalesce(
+                    dayofmonth(dateadd("day", pandas_lit(1), col(column))) == 1,
+                    pandas_lit(False),
+                )
+            ),
         }
         property_function = dt_property_to_function_map.get(property_name)
         if not property_function: